added personService + addPerson api
